requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==2.4.0
constantly==15.1.0
cryptography==3.4.7
daphne==2.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-graphql-jwt==0.3.1
graphene==2.1.8
graphene-django==2.15.0
graphene-subscriptions==1.0.2
graphql-core==2.3.2
graphql-relay==2.0.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pytz==2021.1
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
Twisted==21.2.0
txaio==21.2.1
zope.interface==5.4.0
